Add support for running on AppEngine.
