+++ stack/stack_pool_chan.go
@@ -1,4 +1,4 @@
-// +build !go1.3
+// +build !go1.3 appengine
 
 package stack
